# UniverCity-ui/src/app/Services/models/university-review.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { University } from '../models/university';
import { User } from '../models/user';
export interface UniversityReview {
  createdDate?: string;
  lastModifiedDate?: string;
  rating?: number;
  review?: string;
  university?: University;
  universityReview_id?: number;
  user?: User;
}