# UniverCity-ui/src/openapi/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenApi specification - Shahroz","description":"API Specification for UniverCity","contact":{"name":"Shahroz","email":"shahrozsalman7@gmail.com"},"license":{"name":"Licended by Shahroz"},"version":"1.0"},"servers":[{"url":"http://localhost:8088/api/v1","description":"Local ENV"},{"url":"https://UniverCity.com","description":"PROD ENV"}],"security":[{"bearerAuth":[]}],"paths":{"/api/reviews/editReview/{reviewId}":{"put":{"tags":["university-review-controller"],"operationId":"editReview","parameters":[{"name":"reviewId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UniversityReviewRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UniversityReview"}}}}}}},"/whatsapp/send":{"post":{"tags":["whatsapp-controller"],"operationId":"sendMessage","parameters":[{"name":"to","in":"query","required":true,"schema":{"type":"string"}},{"name":"message","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/university":{"get":{"tags":["University"],"operationId":"findAllUniversity","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResponseUniversityResponse"}}}}}},"post":{"tags":["University"],"operationId":"saveUniversity","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UniversityRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int64"}}}}}}},"/quiz":{"post":{"tags":["Quiz"],"operationId":"saveQuizQuestion","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuizQuestionRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int64"}}}}}}},"/quiz/questionSet":{"post":{"tags":["Quiz"],"operationId":"getQuizQuestions","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuestionSetDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/QuizQuestion"}}}}}}}},"/quiz/questionSetByMain":{"post":{"tags":["Quiz"],"operationId":"getQuizQuestionsByMainCategory","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuestionSetDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/QuizQuestion"}}}}}}}},"/quiz/createUserQuiz":{"post":{"tags":["Quiz"],"operationId":"createUserQuiz","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserQuizDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserQuiz"}}}}}}},"/favorites/{universityId}":{"post":{"tags":["Favourites"],"operationId":"addFavorite","parameters":[{"name":"universityId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}},"delete":{"tags":["Favourites"],"operationId":"removeFavorite","parameters":[{"name":"universityId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/auth/register":{"post":{"tags":["Authentication"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegistrationRequest"}}},"required":true},"responses":{"202":{"description":"Accepted","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/auth/authenticate":{"post":{"tags":["Authentication"],"operationId":"authenticate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticationResponse"}}}}}}},"/api/reviews/addReview/{universityId}/user/{userId}":{"post":{"tags":["university-review-controller"],"operationId":"addReview","parameters":[{"name":"universityId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UniversityReviewRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UniversityReview"}}}}}}},"/user":{"get":{"tags":["User"],"operationId":"getAllUsers","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}}}},"/universityFilter/getCities":{"get":{"tags":["UniversityFilter"],"operationId":"getCities","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CityDTO"}}}}}}}},"/university/{university-id}":{"get":{"tags":["University"],"operationId":"findUniversityById","parameters":[{"name":"university-id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UniversityResponse"}}}}}}},"/university/search/{keyword}":{"get":{"tags":["University"],"operationId":"searchUniversities","parameters":[{"name":"keyword","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/University"}}}}}}}},"/university/getFacilities/{university-id}":{"get":{"tags":["University"],"operationId":"getFacilitiesByUniversity","parameters":[{"name":"university-id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Facility"}}}}}}}},"/university/filter":{"get":{"tags":["University"],"operationId":"filterInstitutions","parameters":[{"name":"filter","in":"query","required":true,"schema":{"$ref":"#/components/schemas/UniversityFilter"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/University"}}}}}}}},"/quiz/getUserQuiz":{"get":{"tags":["Quiz"],"operationId":"getUserQuiz","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserQuiz"}}}}}}}},"/quiz/getSubCategory/{category_id}":{"get":{"tags":["Quiz"],"operationId":"getSubCategoryByCategory","parameters":[{"name":"category_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/QuizSubCategory"}}}}}}}},"/quiz/getQuizBySub/{category_id}/{subcategory_id}":{"get":{"tags":["Quiz"],"operationId":"getQuizBySubCategory","parameters":[{"name":"category_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"subcategory_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/QuizQuestion"}}}}}}}},"/quiz/categories":{"get":{"tags":["Quiz"],"operationId":"getQuizCategories","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/QuizCategory"}}}}}}}},"/favorites":{"get":{"tags":["Favourites"],"operationId":"getFavorites","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/University"}}}}}}}},"/favorites/exists/{universityId}":{"get":{"tags":["Favourites"],"operationId":"isFavorite","parameters":[{"name":"universityId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"boolean"}}}}}}},"/auth/current-user":{"get":{"tags":["Authentication"],"operationId":"getCurrentUser","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/auth/activate-account":{"get":{"tags":["Authentication"],"operationId":"confirm","parameters":[{"name":"token","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/api/reviews/getRating/{universityId}":{"get":{"tags":["university-review-controller"],"operationId":"getUniversityRating","parameters":[{"name":"universityId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"number","format":"double"}}}}}}},"/api/reviews/allReviews/{universityId}":{"get":{"tags":["university-review-controller"],"operationId":"getAllReviews","parameters":[{"name":"universityId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UniversityReview"}}}}}}}},"/api/reviews/deleteReview/{reviewId}":{"delete":{"tags":["university-review-controller"],"operationId":"deleteReview","parameters":[{"name":"reviewId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"UniversityReviewRequest":{"type":"object","properties":{"review":{"type":"string"},"rating":{"type":"integer","format":"int32"},"lastModifiedDate":{"type":"string","format":"date-time"},"createdDate":{"type":"string","format":"date-time"}}},"AcademicScore":{"type":"object","properties":{"academicscore_id":{"type":"integer","format":"int64"},"matricScore":{"type":"number","format":"float"},"intermediateScore":{"type":"number","format":"float"}}},"City":{"type":"object","properties":{"city_id":{"type":"integer","format":"int64"},"name":{"type":"string"},"province":{"type":"string"},"universityLocations":{"type":"array","items":{"$ref":"#/components/schemas/UniversityLocation"}},"userLocations":{"type":"array","items":{"$ref":"#/components/schemas/UserLocation"}}}},"Facility":{"type":"object","properties":{"facility_id":{"type":"integer","format":"int64"},"facilityName":{"type":"string"},"university":{"$ref":"#/components/schemas/University"}}},"GrantedAuthority":{"type":"object","properties":{"authority":{"type":"string"}}},"Notification":{"type":"object","properties":{"notification_id":{"type":"integer","format":"int64"},"message":{"type":"string"},"generatedAt":{"type":"string","format":"date-time"},"alertAt":{"type":"string","format":"date-time"},"read":{"type":"boolean"},"user":{"$ref":"#/components/schemas/User"}}},"Program":{"type":"object","properties":{"program_id":{"type":"integer","format":"int64"},"name":{"type":"string"},"degreeType":{"type":"string"}}},"Role":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"createdDate":{"type":"string","format":"date-time"},"lastModifiedDate":{"type":"string","format":"date-time"}}},"University":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"createdDate":{"type":"string","format":"date-time"},"lastModifiedDate":{"type":"string","format":"date-time"},"createdBy":{"type":"string"},"lastModifiedBy":{"type":"string"},"name":{"type":"string"},"about":{"type":"string"},"rank":{"type":"integer","format":"int32"},"applyDate":{"type":"string","format":"date-time"},"entryTestDate":{"type":"string","format":"date-time"},"startDate":{"type":"string","format":"date-time"},"websiteLink":{"type":"string"},"accrediatetionBody":{"type":"string"},"averageFees":{"type":"number","format":"double"},"universityContacts":{"type":"array","items":{"$ref":"#/components/schemas/UniversityContact"}},"programs":{"type":"array","items":{"$ref":"#/components/schemas/Program"}},"reviews":{"type":"array","items":{"$ref":"#/components/schemas/UniversityReview"}},"facilities":{"type":"array","items":{"$ref":"#/components/schemas/Facility"}},"universityLocations":{"type":"array","items":{"$ref":"#/components/schemas/UniversityLocation"}},"programCount":{"type":"integer","format":"int32"}}},"UniversityContact":{"type":"object","properties":{"universitycontact_id":{"type":"integer","format":"int64"},"email":{"type":"string"},"phoneNumber":{"type":"string"},"university":{"$ref":"#/components/schemas/University"}}},"UniversityLocation":{"type":"object","properties":{"universitylocation_id":{"type":"integer","format":"int64"},"latitude":{"type":"number","format":"float"},"longitude":{"type":"number","format":"float"},"city":{"$ref":"#/components/schemas/City"}}},"UniversityReview":{"type":"object","properties":{"universityReview_id":{"type":"integer","format":"int64"},"createdDate":{"type":"string","format":"date-time"},"lastModifiedDate":{"type":"string","format":"date-time"},"review":{"type":"string"},"rating":{"type":"integer","format":"int32"},"university":{"$ref":"#/components/schemas/University"},"user":{"$ref":"#/components/schemas/User"}}},"User":{"type":"object","properties":{"user_id":{"type":"integer","format":"int64"},"firstName":{"type":"string"},"lastName":{"type":"string"},"phoneNumber":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"accountLocked":{"type":"boolean"},"enabled":{"type":"boolean"},"roles":{"type":"array","items":{"$ref":"#/components/schemas/Role"}},"createdDate":{"type":"string","format":"date-time"},"lastModifiedData":{"type":"string","format":"date-time"},"academicScore":{"$ref":"#/components/schemas/AcademicScore"},"universityreview":{"$ref":"#/components/schemas/UniversityReview"},"notifications":{"type":"array","items":{"$ref":"#/components/schemas/Notification"}},"userLocations":{"type":"array","items":{"$ref":"#/components/schemas/UserLocation"}},"name":{"type":"string"},"username":{"type":"string"},"authorities":{"type":"array","items":{"$ref":"#/components/schemas/GrantedAuthority"}},"credentialsNonExpired":{"type":"boolean"},"accountNonExpired":{"type":"boolean"},"accountNonLocked":{"type":"boolean"}}},"UserLocation":{"type":"object","properties":{"userlocation_id":{"type":"integer","format":"int64"},"longitude":{"type":"string"},"latitude":{"type":"string"},"city":{"$ref":"#/components/schemas/City"},"user":{"$ref":"#/components/schemas/User"}}},"UniversityRequest":{"required":["about","accrediatetionBody","name","rank","websiteLink"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"about":{"type":"string"},"rank":{"type":"integer","format":"int32"},"entryTestDate":{"type":"string","format":"date-time"},"applyDate":{"type":"string","format":"date-time"},"startDate":{"type":"string","format":"date-time"},"websiteLink":{"type":"string"},"accrediatetionBody":{"type":"string"},"averageFees":{"type":"number","format":"double"}}},"QuizQuestionRequest":{"required":["correctAnswer","question"],"type":"object","properties":{"quizquestion_id":{"type":"integer","format":"int64"},"question":{"type":"string"},"option1":{"type":"string"},"option2":{"type":"string"},"option3":{"type":"string"},"option4":{"type":"string"},"correctAnswer":{"type":"string"},"difficultyLevel":{"type":"integer","format":"int32"}}},"QuestionSetDTO":{"type":"object","properties":{"difficulty":{"type":"integer","format":"int32"},"subCategoryId":{"type":"integer","format":"int64"},"categoryId":{"type":"integer","format":"int64"},"count":{"type":"integer","format":"int32"}}},"QuestionSubMain":{"type":"object","properties":{"quizSubCategoryMainCategory":{"$ref":"#/components/schemas/QuizSubCategoryMainCategory"},"questionSubMain_id":{"type":"integer","format":"int64"}}},"QuizCategory":{"type":"object","properties":{"quizcategory_id":{"type":"integer","format":"int64"},"name":{"type":"string"}}},"QuizQuestion":{"type":"object","properties":{"quizquestion_id":{"type":"integer","format":"int64"},"question":{"type":"string"},"option1":{"type":"string"},"option2":{"type":"string"},"option3":{"type":"string"},"option4":{"type":"string"},"correctAnswer":{"type":"string"},"difficultyLevel":{"type":"integer","format":"int32"},"questionSubMains":{"type":"array","items":{"$ref":"#/components/schemas/QuestionSubMain"}}}},"QuizSubCategory":{"type":"object","properties":{"quizsubcategory_id":{"type":"integer","format":"int64"},"name":{"type":"string"}}},"QuizSubCategoryMainCategory":{"type":"object","properties":{"quizCategory":{"$ref":"#/components/schemas/QuizCategory"},"quizSubCategory":{"$ref":"#/components/schemas/QuizSubCategory"},"quizSubCategoryMainCategory_id":{"type":"integer","format":"int64"}}},"UserQuizDTO":{"type":"object","properties":{"score":{"type":"string"},"completed":{"type":"number","format":"float"},"categoryId":{"type":"integer","format":"int64"},"time":{"type":"string"}}},"UserQuiz":{"type":"object","properties":{"userquiz_id":{"type":"integer","format":"int64"},"date":{"type":"string","format":"date"},"score":{"type":"string"},"completed":{"type":"number","format":"float"},"time":{"type":"string"},"quizCategory":{"$ref":"#/components/schemas/QuizCategory"},"user":{"$ref":"#/components/schemas/User"}}},"RegistrationRequest":{"required":["email","firstName","lastName","password","phoneNumber"],"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"phoneNumber":{"type":"string"},"email":{"type":"string"},"password":{"maxLength":2147483647,"minLength":8,"type":"string"}}},"AuthenticationRequest":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"maxLength":2147483647,"minLength":8,"type":"string"}}},"AuthenticationResponse":{"type":"object","properties":{"token":{"type":"string"}}},"PageResponseUniversityResponse":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/UniversityResponse"}},"number":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"}}},"UniversityResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"about":{"type":"string"},"rank":{"type":"integer","format":"int32"},"applyDate":{"type":"string","format":"date-time"},"startDate":{"type":"string","format":"date-time"},"entryTestDate":{"type":"string","format":"date-time"},"websiteLink":{"type":"string"},"accrediatetionBody":{"type":"string"},"averageFees":{"type":"number","format":"double"}}},"CityDTO":{"type":"object","properties":{"name":{"type":"string"},"province":{"type":"string"}}},"UniversityFilter":{"type":"object","properties":{"cities":{"type":"array","items":{"type":"string"}},"minFees":{"type":"number","format":"double"},"maxFees":{"type":"number","format":"double"},"accreditationBodies":{"type":"array","items":{"type":"string"}},"sectors":{"type":"array","items":{"type":"string"}},"startDate":{"type":"string","format":"date"},"endDate":{"type":"string","format":"date"},"program":{"type":"array","items":{"type":"string"}}}}},"securitySchemes":{"bearerAuth":{"type":"http","description":"JWT Auth Description","in":"header","scheme":"bearer","bearerFormat":"JWT"}}}}